WIP rockpi/s: new configuration
Add configuration package for the RockPi S board.

TODO:

 - SPL / TPL build process
 - http://opensource.rock-chips.com/wiki_Boot_option

Currently the u-boot cannot build SPL or TPL for this board. This leaves us with
a requirement to use the closed-source tools and binary blobs provided in
the "rockchip binary repository." This is somewhat unacceptable as we generally
do not use closed-source host tools in Buildroot (although proprietary
bootloader and firmware blobs are common).